requirements.txt
flask
torch
torchvision
numpy
diffusers[torch]
transformers
Pillow